Create CleanHighlightMarkDown for use in help-related searches. Style updates.
