+++ src/help/components/landing/HelpAutocomplete.tsx
@@ -1,15 +1,13 @@
-// TODO: fix import order
 import { ChangeEvent, useState } from 'react';
-import { Card, DataList, SearchInput } from 'franklin-sites';
+import { Card, InfoList, SearchInput } from 'franklin-sites';
 
 import useDataApi from '../../../shared/hooks/useDataApi';
 
-import HelpCompactResult from '../shared/HelpCompactResult';
+import CleanHighlightMarkDown from '../../utils/CleanHighlightMarkDown';
 
 import { help as helpURL } from '../../../shared/config/apiUrls';
 
 import { HelpSearchResponse } from '../../adapters/helpConverter';
-import { getIdKey } from '../results/Results';
 
 import styles from './styles/help-autocomplete.module.scss';
 
@@ -18,6 +16,19 @@ const HelpAutocomplete = () => {
   const dataObject = useDataApi<HelpSearchResponse>(
     searchValue && helpURL.search({ query: searchValue })
   );
+
+  const articles = dataObject?.data?.results.map((article) => {
+    const titleMatch = article.matches?.title?.[0];
+    const contentMatch = article.matches?.content?.[0];
+    return {
+      title: titleMatch ? (
+        <CleanHighlightMarkDown md={titleMatch} />
+      ) : (
+        article.title
+      ),
+      content: contentMatch && <CleanHighlightMarkDown md={contentMatch} />,
+    };
+  });
 
   return (
     <div className={styles['help-autocomplete']}>
@@ -30,20 +41,9 @@ const HelpAutocomplete = () => {
         value={searchValue}
         autoFocus
       />
-      {!!dataObject?.data?.results.length && (
+      {!!articles?.length && (
         <Card>
-          <DataList
-            getIdKey={getIdKey}
-            data={dataObject.data.results}
-            dataRenderer={(article) => (
-              <HelpCompactResult
-                id={article.id}
-                title={article.title}
-                titleMatch={article.matches?.title?.[0]}
-                contentMatch={article.matches?.content?.[0]}
-              />
-            )}
-          />
+          <InfoList infoData={articles} />
         </Card>
       )}
     </div>
