+++ src/help/components/landing/HelpLandingPage.tsx
@@ -1,6 +1,4 @@
-// TODO: fix import order
 import HelpAutocomplete from './HelpAutocomplete';
-// import styles from './styles/help-landing-page.module.scss';
 
 const HelpLandingPage = () => (
   <div className="uniprot-grid--centered">
