+++ src/help/components/landing/styles/help-autocomplete.module.scss
@@ -2,9 +2,18 @@
   :global(input[type='text']) {
     margin-bottom: 0.5rem;
   }
+
   :global(.card) {
     // remove the fake margin on the right of the cards
     width: initial;
     margin: 0;
+    max-height: 30rem;
+    overflow-y: scroll;
+  }
+
+  :global(.decorated-list-item__title) {
+    flex-basis: 30vw;
+    text-align: left;
+    border-right-color: #d2dce3;
   }
 }
