+++ src/help/components/results/HelpCard.tsx
@@ -1,23 +1,12 @@
 import { memo } from 'react';
 import { generatePath, Link } from 'react-router-dom';
 import { Card } from 'franklin-sites';
-import marked from 'marked';
-
-import cleanText, {
-  cleanTextDefaultOptions,
-} from '../../../shared/utils/cleanText';
 
 import { LocationToPath, Location } from '../../../app/config/urls';
 
-import styles from './styles/help-card.module.scss';
+import CleanHighlightMarkDown from '../../utils/CleanHighlightMarkDown';
 
-const cleanTextOptions = {
-  ...cleanTextDefaultOptions,
-  allowedClasses: {
-    ...cleanTextDefaultOptions.allowedClasses,
-    '*': ['match-highlight'],
-  },
-};
+import '../../utils/highlight.scss';
 
 type Props = {
   id: string;
@@ -36,37 +25,14 @@ const HelpCard = ({ id, title, titleMatch, contentMatch }: Props) => {
       header={
         <h2 className="tiny">
           <Link to={to}>
-            {titleMatch ? (
-              <span
-                // eslint-disable-next-line react/no-danger
-                dangerouslySetInnerHTML={{
-                  __html: cleanText(
-                    marked.parseInline(titleMatch),
-                    cleanTextOptions
-                  ),
-                }}
-              />
-            ) : (
-              title
-            )}
+            {titleMatch ? <CleanHighlightMarkDown md={titleMatch} /> : title}
           </Link>
         </h2>
       }
       headerSeparator={false}
       to={to}
-      className={styles['help-card']}
     >
-      {contentMatch && (
-        <div
-          // eslint-disable-next-line react/no-danger
-          dangerouslySetInnerHTML={{
-            __html: cleanText(
-              marked.parseInline(contentMatch),
-              cleanTextOptions
-            ),
-          }}
-        />
-      )}
+      {contentMatch && <CleanHighlightMarkDown md={contentMatch} />}
     </Card>
   );
 };
