--- src/help/components/shared/HelpCompactResult.tsx
@@ -1,14 +1,0 @@
-type Props = {
-  id: string;
-  title: string;
-  titleMatch?: string;
-  contentMatch?: string;
-};
-
-const HelpCompactResult = ({ id, title, titleMatch, contentMatch }: Props) => (
-  <>
-    {id} {title} {titleMatch} {contentMatch}
-  </>
-);
-
-export default HelpCompactResult;
