+++ src/help/utils/highlight.scss
@@ -1,6 +1,6 @@
 @import 'franklin-sites/src/styles/colours';
 
-:global(.match-highlight) {
+.match-highlight {
   position: relative;
   z-index: 0;
